requirements.txt
torch>=2.1
torchvision>=0.16
opencv-python
ultralytics
numpy
flask
requests
pyyaml

